Tied Scores == Tied Ranks
When Raw Score and/or Tier scores are tied and calculate is selected,
ranks given to the tied teams will be the same. (Rank count will
continue to increase to make room for the tied teams) Ex. If Two teams
are tied for 5th place, both would temporarily receive rank 5. the next
team would be given rank 7.